Merge WalletManager into Wallet
Parts are being moved around to accomodate for the backend endpoints and what it needs. Wallet has been moved closer to a one to one mapping
with user. There are some almost circular dependencies with wallet passwords and private keys that I am trying to ensure don't get too entangled.