Replace Wallet with HDWallet
Instead of generating a wallet from private key generate a HD wallet from seed. From the seed generate two nodes.
One node is for receiving funds, one is for identity. These nodes can be recreated from the seed.
The seed is stored to SharedPreferences only (it is no longer encrypted and stored on the backend).