Move Contacts list into child fragment
Top level fragments are just containers for all child fragments. That means the activity contains a holder for fragments that are containers for child fragments. It's fragments all the way down; let's hope this approach works. On the plus side logic is outside of the fragments in presenters so I am feeling more confident that this can work.