Pass absent Optional instead of null in message sender
An Optional.<SignalServiceMessageSender.EventListener> is
expected by SignalServiceMessageSender, but passing null
instead eventually results in a call to null.isPresent()
during encryption of outgoing message, causing prekey
fetching to fail. In the case of the first message sent
to a recipient, this is fatal, as no prekeys are cached.

Also fixed is a IllegalStateException thrown during
receiving of a message, caused by get()ing an absent
Optional.