Major fix: Realm migration didn't work
The migration to rename owner_address to token_id doesn't work on many devices for reasons I am unsure of. This change
removes the migration and just keeps the name the same. Unfortunately devs may have to reinstall the app their running if their realm is in a weird state.