Don't crash when waking app via PN
Clicking a push notification when the app was dead would crash. This is because we have two step initialisation and the databases were
being referenced before they existed.

Luckily, initialising a database no longer has any consequences because the databases are thread safe!
Moving the creation of the stores into the constructor has no costs and stops the app crashing.