Add offline snackbar.
Create an OfflineViewBasePresenterActivity that an Activity can choose to extend; by doing so the Activity needs to provide
a view in which to render any eventual offline snackbar. This allows the Activity tehmselves to have control over whether or
not an offline snackbar should be rendered whilst keeping the offline logic in one centralised class.