Delay adding new messages until conversation loaded.
New messages could be added before a conversation was loaded. This caused an unusual race condition.

A new message would be added, then the conversation would load and override that new message from bring shown.
Interestingly, when the new message went from "sending" to "sent" the new message would render; but we don't scroll when updating messages.

This lead to a situation where an "updated" message was being added to the adapter but the adapter wasn't scrolling.
This made it look like scrolling was broken when in fact it wasn't; there was a nasty race condition hiding the bug.