Bug fix: Remove ReplaySubject to suppress double message sending
ReplaySubject was being used as a queue to queue up messages that couldn't be sent yet.
Unfortunately that meant on configuration change the subject was being subscribed too again and thus replaying all the messages.
This leads to messages being sent several times.

Instead of using the subject as a queue, I use a list as a queue and broadcast that list to the subject when it becomes connected.
The logic here is growing unwiedly and it would be nice to start pulling code out of ChatPresenter.